Set default min days.

+++ transitclock/src/main/java/org/transitclock/core/predictiongenerator/kalman/KalmanPredictionGeneratorImpl.java
@@ -43,7 +43,7 @@ public class KalmanPredictionGeneratorImpl extends HistoricalAveragePredictionGe
 	 * historical value. 
 	 */
 	private static final IntegerConfigValue minKalmanDays = new IntegerConfigValue(
-			"transitclock.prediction.data.kalman.mindays", new Integer(1),
+			"transitclock.prediction.data.kalman.mindays", new Integer(3),
 			"Min number of days trip data that needs to be available before Kalman prediciton is used instead of default transiTime prediction.");
 
 	private static final IntegerConfigValue maxKalmanDays = new IntegerConfigValue(
